#87: moved assets to webui folder
