+++ src/webui/components/application/Introduction.tsx
@@ -6,7 +6,7 @@ import { Button, Column, Image, Panel, Row, Text } from '../../designsystem/modu
 
 import { ApplicationLogo } from '../module';
 
-import introductionImage from '../../../assets/images/introduction.png';
+import introductionImage from '../../assets/images/introduction.png';
 
 export default function Feature()
 {
