+++ src/webui/components/application/Logo.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { Image } from '../../designsystem/module';
 
-import logoImage from '../../../assets/images/logo.svg';
+import logoImage from '../../assets/images/logo.svg';
 
 export default function Component()
 {
