+++ src/webui/components/comic/Image.tsx
@@ -5,7 +5,7 @@ import { Image } from '../../designsystem/module';
 
 import type ComicView from '../../../domain/comic/view/ComicView';
 
-import dummyDataUrl from '../../../assets/images/introduction.png';
+import dummyDataUrl from '../../assets/images/introduction.png';
 
 export type Props = {
     comic: ComicView;
